+++ src/pages/staking/staking.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef, useState } from 'react';
 import './staking.less';
 import 'antd/dist/antd.css';
-import { AutoComplete, Button, Form, Input, InputNumber, Layout, Table, Tabs } from 'antd';
-// import {ReactComponent as HomeIcon} from '../../assets/icon-home-white.svg';
+import { Button, Checkbox, Form, Input, InputNumber, Layout, Table, Tabs } from 'antd';
+import { OrderedListOutlined } from '@ant-design/icons';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { AddressType } from '@crypto-com/chain-jslib/lib/dist/utils/address';
 import Big from 'big.js';
@@ -12,9 +12,9 @@ import SuccessModalPopup from '../../components/SuccessModalPopup/SuccessModalPo
 import ErrorModalPopup from '../../components/ErrorModalPopup/ErrorModalPopup';
 import PasswordFormModal from '../../components/PasswordForm/PasswordFormModal';
 import { secretStoreService } from '../../storage/SecretStoreService';
-import { sessionState, validatorTopListState, walletAssetState } from '../../recoil/atom';
+import { sessionState, walletAssetState } from '../../recoil/atom';
 import { scaledBalance, UserAsset } from '../../models/UserAsset';
-import { BroadCastResult, RewardTransaction } from '../../models/Transaction';
+import { BroadCastResult, RewardTransaction, ValidatorModel } from '../../models/Transaction';
 import { TransactionUtils } from '../../utils/TransactionUtils';
 import { FIXED_DEFAULT_FEE } from '../../config/StaticConfig';
 import {
@@ -23,8 +23,10 @@ import {
   getCurrentMinAssetAmount,
   getUIDynamicAmount,
 } from '../../utils/NumberUtils';
+import { middleEllipsis } from '../../utils/utils';
 
 const { Header, Content, Footer } = Layout;
+const { Search } = Input;
 const { TabPane } = Tabs;
 const layout = {
   // labelCol: { span: 8 },
@@ -55,22 +57,34 @@ const FormDelegationRequest = () => {
   const [errorMessages, setErrorMessages] = useState([]);
   const [inputPasswordVisible, setInputPasswordVisible] = useState(false);
   const [decryptedPhrase, setDecryptedPhrase] = useState('');
+  const [isValidatorListVisible, setIsValidatorListVisible] = useState(false);
+  const [showMemo, setShowMemo] = useState(false);
   const [walletAsset, setWalletAsset] = useRecoilState(walletAssetState);
   const currentSession = useRecoilValue(sessionState);
-  const [validatorTopList, setValidatorTopList] = useRecoilState(validatorTopListState);
+  const [validatorTopList, setValidatorTopList] = useState<ValidatorModel[]>([]);
   const didMountRef = useRef(false);
+
+  const processValidatorList = (validatorList: ValidatorModel[]) => {
+    return validatorList.map((validator, idx) => {
+      const validatorModel = {
+        ...validator,
+        key: `${idx}`,
+      };
+      return validatorModel;
+    });
+  };
 
   useEffect(() => {
     let unmounted = false;
 
     const syncValidatorsData = async () => {
-      const currentValidatorList =
-        validatorTopList.length === 0
-          ? await walletService.getLatestTopValidators()
-          : validatorTopList;
+      const currentValidatorList = await walletService.retrieveTopValidators(
+        currentSession.wallet.identifier,
+      );
 
       if (!unmounted) {
-        setValidatorTopList(currentValidatorList);
+        const validatorList = processValidatorList(currentValidatorList);
+        setValidatorTopList(validatorList);
       }
     };
 
@@ -82,7 +96,7 @@ const FormDelegationRequest = () => {
     return () => {
       unmounted = true;
     };
-  }, [validatorTopList]);
+  }, [validatorTopList, setValidatorTopList]);
 
   const showConfirmationModal = () => {
     setInputPasswordVisible(false);
@@ -192,6 +206,76 @@ const FormDelegationRequest = () => {
     )} ${walletAsset.symbol}`,
   );
 
+  const validatorColumns = [
+    {
+      title: 'Name',
+      dataIndex: 'validatorName',
+      key: 'validatorName',
+    },
+    {
+      title: 'Domain',
+      dataIndex: 'validatorWebSite',
+      key: 'validatorWebSite',
+      render: validatorWebSite => {
+        return validatorWebSite === '' ? (
+          'n.a.'
+        ) : (
+          <a
+            data-original={validatorWebSite}
+            target="_blank"
+            rel="noreferrer"
+            href={`${validatorWebSite}`}
+          >
+            {validatorWebSite}
+          </a>
+        );
+      },
+    },
+    {
+      title: 'Address',
+      dataIndex: 'validatorAddress',
+      key: 'validatorAddress',
+      render: validatorAddress => (
+        <a
+          data-original={validatorAddress}
+          target="_blank"
+          rel="noreferrer"
+          href={`${currentSession.wallet.config.explorerUrl}/validator/${validatorAddress}`}
+        >
+          {middleEllipsis(validatorAddress, 12)}
+        </a>
+      ),
+    },
+    {
+      title: 'Commission Rate',
+      dataIndex: 'currentCommissionRate',
+      key: 'currentCommissionRate',
+      render: currentCommissionRate => (
+        <span>{new Big(currentCommissionRate).times(100).toFixed(2)}%</span>
+      ),
+    },
+    {
+      title: 'Action',
+      key: 'action',
+      render: record => (
+        <a
+          onClick={() => {
+            setIsValidatorListVisible(false);
+            form.setFieldsValue({
+              validatorAddress: record.validatorAddress,
+            });
+          }}
+        >
+          Select
+        </a>
+      ),
+    },
+  ];
+
+  function onShowMemoChange() {
+    setShowMemo(!showMemo);
+  }
+
   return (
     <Form
       {...layout}
@@ -201,6 +285,27 @@ const FormDelegationRequest = () => {
       onFinish={showPasswordInput}
       requiredMark={false}
     >
+      <>
+        <ModalPopup
+          isModalVisible={isValidatorListVisible}
+          handleCancel={() => setIsValidatorListVisible(false)}
+          handleOk={() => setIsValidatorListVisible(false)}
+          className="validator-modal"
+          footer={[]}
+          okText="Confirm"
+          width={1000}
+        >
+          <div className="title">Validator List</div>
+          <div className="description">Please select one of the validator.</div>
+          <div className="item">
+            <Table
+              dataSource={validatorTopList}
+              columns={validatorColumns}
+              pagination={{ showSizeChanger: false }}
+            />
+          </div>
+        </ModalPopup>
+      </>
       <Form.Item
         name="validatorAddress"
         label="Validator address"
@@ -210,19 +315,12 @@ const FormDelegationRequest = () => {
           { required: true, message: 'Validator address is required' },
           customAddressValidator,
         ]}
+        className="input-validator-address"
       >
-        <AutoComplete
-          options={[
-            {
-              label: 'Top Validators',
-              options: validatorTopList.map(e => {
-                return {
-                  value: e.validatorAddress,
-                };
-              }),
-            },
-          ]}
+        <Search
           placeholder="Enter validator address"
+          enterButton={<OrderedListOutlined />}
+          onSearch={() => setIsValidatorListVisible(true)}
         />
       </Form.Item>
       <div className="amount">
@@ -251,9 +349,18 @@ const FormDelegationRequest = () => {
           </div>
         </div>
       </div>
-      <Form.Item name="memo" label="Memo (Optional)">
-        <Input />
-      </Form.Item>
+      <Checkbox onChange={onShowMemoChange} checked={showMemo}>
+        Want to set custom memo?
+      </Checkbox>
+      {showMemo ? (
+        <div style={{ paddingTop: '12px' }}>
+          <Form.Item name="memo" label="Memo (Optional)">
+            <Input />
+          </Form.Item>
+        </div>
+      ) : (
+        <div />
+      )}
       <Form.Item {...tailLayout}>
         <ModalPopup
           isModalVisible={isConfirmationModalVisible}