+++ src/recoil/atom.tsx
@@ -2,8 +2,7 @@ import { atom } from 'recoil';
 import { DefaultWalletConfigs } from '../config/StaticConfig';
 import { Session } from '../models/Session';
 import { Wallet } from '../models/Wallet';
-import { UserAsset } from '../models/UserAsset';
-import { ValidatorModel } from '../models/Transaction';
+import { UserAsset, AssetMarketPrice } from '../models/UserAsset';
 import { NORMAL_WALLET_TYPE } from '../service/LedgerService';
 
 const wallet = new Wallet(
@@ -29,6 +28,12 @@ const asset: UserAsset = {
   description: 'Default Asset',
   decimals: 1,
 };
+const market: AssetMarketPrice = {
+  assetSymbol: 'CRO',
+  currency: 'USD',
+  dailyChange: '+0.00',
+  price: '0.000',
+};
 
 const walletIdentifierState = atom({
   key: 'walletIdentifier',
@@ -38,6 +43,11 @@ const walletIdentifierState = atom({
 const sessionState = atom({
   key: 'session',
   default: session,
+});
+
+const marketState = atom({
+  key: 'market',
+  default: market,
 });
 
 const walletAssetState = atom({
@@ -56,11 +66,6 @@ const walletTempBackupState = atom<Wallet | null>({
   default: null,
 });
 
-const validatorTopListState = atom<ValidatorModel[]>({
-  key: 'validatorTopList',
-  default: [],
-});
-
 const hasShownWarningOnWalletTypeState = atom<boolean>({
   key: 'hasShownWarningOnWalletTypeState',
   default: false,
@@ -69,9 +74,9 @@ const hasShownWarningOnWalletTypeState = atom<boolean>({
 export {
   walletIdentifierState,
   sessionState,
+  marketState,
   walletAssetState,
   walletListState,
   walletTempBackupState,
-  validatorTopListState,
   hasShownWarningOnWalletTypeState,
 };