+++ CHANGELOG.md
@@ -6,6 +6,19 @@ All notable changes to this project will be documented in this file.
 
 *Released*
 
+## [v0.1.3] - 2021-04-01
+
+### Additions
+- A newly revamped validator pop up screen selection
+- Now loading up to 20 top validators for selection
+- Show assets values in USD price on home screen
+- Show more informative notice on redelegate confirmations
+- Append default client desktop memo when no memo is provided
+
+### Bug Fixes
+- Removed the extra text and space that was being added when address is copied
+- Fixed misleading warning before wallet deletion for Ledger wallets
+
 ## [v0.1.2] - 2021-03-29
 
 ### Bug Fixes